Added a bunch of fonts, images for final-ish UI

+++ example/RTC_Driver.c
@@ -111,7 +111,7 @@ void RTCinit(uint8_t hour,uint8_t minute, uint8_t second){
 	rtc_init = 1;
 }
 
-void getTime(int *hour,int *minute){
+void getTime(short *hour,short *minute){
 	if(rtc_init){
 		RTC_GetTime(RTC_Format_BIN, &RTC_TimeStructure);
 		*hour = RTC_TimeStructure.RTC_Hours;
